NOTE: Packages that referenced \@azure/storage-blob needed to reference the current version in master (which is a preview). This should be a temporary issue, but the current state of opentelemetry makes it difficult for us to do a partial migration.
This affects:
* storage-blob-changefeed
* storage-file-datalake
* perf-tests/storage-blob
* quantum-jobs

Mostly straightforward. This is the only code right now that actually deals
with creating spans (that's not test code). Biggest issue is that we have
a wrapper so we can handle OpenCensus which had to map from SpanStatusCode to
CanonicalCode, which is a much broader and more specific case. For now, our translation
just assumes CanonicalCode.INTERNAL for errors.

NOTE: Rev'ing our GLOBAL_TRACER_VERSION, after talking with @xirzec.